requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
pandas==2.1.4
rapidfuzz==3.6.1
email-validator==2.1.0
python-dateutil==2.8.2
pydantic==2.5.2
flask-migrate==4.0.5
marshmallow==3.20.1